Fix ML2 test extension driver API test cases
The test cases were asserting bad behavior by the way the
test driver was designed. Rather than ensuring that the extension
was extending the resource in each case with the correct value,
it was asserting that different values came back from
update/create/show. In other words, it was asserting API
inconsistency which interferred with the dependent patch.

This patch fixes that by having the extension store the values requested
in the API and then use those same values for extending the result
whether it be for updates, creates, or shows.

Change-Id: Ib1517102e3fb776fca818acfddd9e0f733563fb3
Partial-Bug: #1625981
(cherry picked from commit 7e822960425326f26806bbfb650b8a844f38557d)