Stop oslo_messaging from error logging CallbackNotFound
This is expected when the server is pushing for a resource
and some listeners (like the OVS agent) are not interested
in the event. This happens when Trunk create/delete events
are dispatched by the server.

Change-Id: I90ddffda546af45ca85b3e4f4f22eed005d971df
Closes-bug: #1625872
(cherry picked from commit bad3eaace1a320e33d3bcdff43e4c66ff0c49066)