Update trunk metadata during wire/unwire operations
Before this patch metadata on the OVS trunk bridge were
not updated during unwiring operations and upon subsequent
wire operations old subports were getting wiped from the
external_ids field of the OVS parent port. This could lead
to OVS ports left behind on the host on trunk removal.

This patch ensures that the subports stored as metadata
reflect the logical state of the trunk.

Closes-bug: #1625875

Change-Id: I9d885b10d84fad43c527d3ca7bc9a37439197fc9