Don't swallow ImportError from load_class_by_alias_or_classname
Before the patch, we were silently swallowing the original ImportError
raised by load_class_by_alias_or_classname that could contain valuable
details about the reason of failure.

This patch reraises the original exception instead of replacing it with
the one with no context. It still logs an error before doing it to make
sure the operator will see it irrespective of what the caller will do
with the exception.

Also, a note to fellow reviewer: No, I don't think it needs a unit test
case. Do you?

Change-Id: Ib47b67da1e7e4b749abb8b1b93888e23d00da00a