Only mark ports ready on synced networks
The DHCP agent was previously resending every single port to
the server whenever sync_state was called, even if it was just
for one network.

This let to sending way too much unnecessary data to the server
and also potentially resulted in sending a port to the server
that wasn't actually provisioned yet.

This patch corrects the behavior by only sending ports for networks
that are being synced if it's a conditional sync.

Closes-Bug: #1639086
Change-Id: Ie7686837b18ff251baa315ef95dc511cda475672