Add db_add to OVSDB API
It is at times useful to append to an OVSDB row in a single
transaction instead of doing separate transactions to pull back a
value, then another to modify and append to it. This patch adds

API.db_add(self, table, record, column, *values)

to the OVSDB API and adds it to the vsctl and native
implementations. It follows the format of ovs-vsctl's 'add'
command which works on both sets and maps. For map columns, its
behavior is to set a key in a column if it is not already set. For
sets, it appends to the current values stored in the column.

For maps, values can be passed as
  (..., {'key': value}, {'key2': value2, ...)
or
 (..., {'key': value, 'key2': value})

Change-Id: Ib6ce75846f9b13c1c33f0ced5ccc619ee7860dc1