Fix IPv6 PD with pluggable IPAM
A PD-sepcific check that was present in the non-pluggable IPAM
backend was missing from the pluggable version, causing router
interfaces to be created with a '::1' IP (equal to the gateway on
PD subnets). The resulting command:

ip netns exec <router_namespace> ip -6 addr add ::1/64

would then fail. This patch restores the check to ensure that
'::1' is not used for router interface ports, and introduces a
test to protect against future regressions.

Change-Id: I6a2e7cd60dd42d92b900a57bd8f3ffbc5909908e
Closes-Bug: 1631693