Disable 'accept_ra' in DHCP agent namespace
Currently the DHCP agent relies on the acceptance of an
RA to configure its IPv6 address with SLAAC or DHCPv6-stateless
network modes.  It should explicitly assign addresses to the
agent based on the data model instead.

In order to do this we must disable RAs in the namespace so
that a static assignment doesn't conflict with a previously
created dynamically-generated address.

Conflicts:
	neutron/agent/linux/interface.py

Change-Id: I1b38d131249d59fa486a07024d4b1ec61e693d59
Related-bug: #1627902
(cherry picked from commit 904f85e2f933493d386f463af3ff6168dd36b7de)