Pecan: Bulk create with one item returns plural
If a bulk create happened with a single item, the pecan code would
return back to the user the singular version of the resource in the
body.  For example, to bulk create many security group rules, the user
would give in the body of the request the json with a parent key of
"security-group-rules" with a value of a list of json security group
rules.  If this list is of length one, then after the creation of the
one security group rule, the API would return back to the client
"security-group-rule" as the parent key.  This is not how the legacy
wsgi layer behaved.  The behavior is expected when nova creates security
group rules by calling the neutron API through its deprecated security
groups API.

Change-Id: I8757630403e4d486cd3c8dd6f041e9ee326ba3b4
Closes-Bug: #1633671