DHCP agent: advertise SLAAC prefixes
Advertise IPv6 SLAAC prefixes via router advertisements with router
lifetime of zero so prefix can be discovered on isolated networks.
Since the router lifetime of these RAs is zero, they will not be used as
a next hop. Implementing this now so Octavia can use an isolated IPv6
network.

Change-Id: I4a9c2018606d1729465612206f8e8e2fc4922516
Closes-bug: #1498987
Related-bug: #1407573