Reduce rally executions of create_and_list_ports
The previous configuration of the task was taking up
to a half hour to run. Between this task and the others,
it was eating up all of our gate time, leaving no room
to add new jobs.

This reduces it 5 times in the number of runs, from 40
to 8. This still gives us a reasonable number to get an
average from, especially since each run creates 100 ports.

Change-Id: I0955e44df1a9e072c58fdacc337121b8621132df