Introduce objects_exist method in neutron/objects/base.py
This patch introduces objects_exist method in neutron/objects/base.py.

The purpose of this patch: During OVO integration, there are
instances where we need to check whether an object is present
or not. Instead of writing custom methods based on get_objects,
this objects_exist method can be used.

Change-Id: I2546c31ef4a81c1e34dfaade97f93407910e0985
Partially-Implements: blueprint adopt-oslo-versioned-objects-for-db