Use callbacks to create DVR floating GW port
This stops the DVR mixin from overriding the '_update_fip_assoc'
function in order to create its agent floating IP gateway ports.
It instead now subscribes to the floating IP AFTER_UPDATE event
to create these ports.

Additionally, it fixes the semantics of the floating IP AFTER_UPDATE
event so it's emitted after the update is committed to the DB
so it's safe to call core plugin mutation methods and external systems.

In addition to fixing the callback semantics, this is part of the
effort to eliminate GUARD_TRANSACTION in-tree, which
should pave the way for ML2 to adopt the new engine facade and for
push notifications to safely transmit the latest DB state.

Related-Bug: #1540844
Partially-Implements: blueprint push-notifications
Partially-Implements: blueprint enginefacade-switch
Change-Id: I93e4d847f96707b17c4a7dfdb3bbf81d062fe3fb