Document how to proceed with new platform features
Updated effective neutron guide to reflect that we should not start
using new features in code without checking common platform for these
features' availability in distro packages. If some are indeed missing a
feature, we may need to revisit the direction, or talk to distros, or
handle failures gracefully.

Change-Id: I38cbe19ae0134fb129aad1fa80cc712a17ab94ad
Related-Bug: #1644312