Elimination duplication of lots of DVR tests
The L3DvrTestCaseMigration class was just inheriting
and subsequently running all of the tests in the
L3DvrTestCase class. It doesn't have a separate setup
or mixin from the parent class so it was just duplicating
all of the tests for no reason.

Create a "base" test class so the migration class
can just run a single test.

Change-Id: Idbe347fa26e64c600cf3056d16a99c4ad46a2586