Get rid of DVR override of _port_has_ipv6_address
This function was added as part of [1] as a way to
prevent _find_ipv6_router_port_by_network from returning
the DVR SNAT port. However, the same thing can be
achieved by making _find_ipv6_router_port_by_network
just check that the interface matches the expected
device owner for regular router interfaces before
returning it.

This function is the last inherited function in the
DVR mixin other than _get_device_owner.
Once a solution for _get_device_owner and the
router extra attributes processing is found, all of
this code can be moved out of the mixin in to the
DVR driver.

1. Id243a4b3f30071226411ace6d12550fc099901cc

Change-Id: I9923ac41914dcf2e6f047517215e4c1c2d754db3