Get rid of create_router override in l3_ha mixin
Switch to callback events for the create router processing
of HA routers. This maintains the same transactional
semantics as before. However, if an error is encountered
after creating the router during scheduling, it is no
longer deleted but it is instead put into the ERROR state.

Related: blueprint multi-l3-backends
Change-Id: Ib7444408286841e0e3dfdc76339652ceb2e6db94