Redirect ExtensionDescriptor to neutron-lib
Preliminary step in adopting ExtensionDescriptor from neutron-lib.

This switches to using the ExtensionDescriptor class from
neutron-lib. In this step we provide a redirector so that subprojects
do not break. In follow-up patches, all subprojects will switch to
using neutron-lib directly, and then the redirector will be removed
from neutron core.

Change-Id: I5a111e0033e518e39b3042f047ee9eebba77a0d5