Migrate device_owner for router's interface
When migrating a HA router to legacy, the router's interface (whose
device_owner is network:ha_router_replicated_interface) isn't changed
to network:router_interface. This means that certain actions (such as
trying to use 'neutron router-interface-delete' and migrating the router
to DVR) won't work (since a router interface of the appropriate type
can't be found and an SNAT port won't be created, respectively). This
patch takes care of the needed migration.

Closes-Bug: #1645716
Change-Id: I0c18ccf08ddfa57fde9b5ac331101497079bd330