Add filters support to constant queries test
This allows the tests to ensure constant query behavior is observed
when passing in filters. This patch updates the port query to test
with a few different column matches.

This also fixes the test to not observe queries emitted by the quota
engine that happen as a result of a list operation right after a
create operation.

TrivialFix
Change-Id: Ie1eab034ad99a233888cb048319bc31107be6fa3