Optimize trunk details extension hook
As per findings [1], the hook was found to be a burden. That is
because of the lazy join on every DB lookup required to fetch
the MAC for the subport. Instead of issuing a gazillion JOINs
(one per each subport), this patch modifies the loop and speed
it up at the expenses of some temporary memory cost: it does so
by building a temporary map, looking up ports by IDs (which is
bleeding fast) and reassembling the data from the temp map as
the API expects it to be. Pronto!

[1] https://wiki.openstack.org/wiki/Neutron_Trunk_API_Performance_and_Scaling#surprise_effect_on_filtered_port_listings

Closes-bug: #bug/1648714

Change-Id: I58e39dbf1c83e1204d8d7f436395531ca096b1ab