Adding debug trace to IPWrapper.get_devices()
This patch adds a debug trace to the get_devices() method due to some
jobs failing in gate (especifically functional test for netns_cleanup).
From gate logs, looks like rootwrap_daemon is returning the output of
the previous command and this trace is to make sure what the actual
output is.

Change-Id: I83c59b9f0ae1ab51f166f442d99a1022e88d808e
Related-Bug: #1654287