Decompose router extra attr processing
This adjusts the extra router attributes mixin to add
a new function to explicitly set the value of an extra
attribute. The mixins that depend on extra attrs
now explicitly call this instead of depending on a big
inheritance chain of calling _process_extra_attr_router_create.

This will allow the router_az, DVR and HA modules to eventually
be separated from the L3 plugin class inheritance entirely.

Related: blueprint multi-l3-backends
Change-Id: If44bfef9428f43dd82758928bf46786b70e5e11c