Protect against non-determinstic sort
With the switch to subquery loading, we need to eliminate the
possibility for non-deterministic results in the event of LIMIT
queries caused by non-unique SORT BY parameters.

This always adds on a set of unique keys to sort by when sorting
params are present. By placing them at the end, they take the lower
priority over the user-provided sort strings.

1. http://docs.sqlalchemy.org/en/latest/orm/loading_relationships.html#subqueryload-ordering
2. http://docs.sqlalchemy.org/en/latest/faq/ormconfiguration.html#faq-subqueryload-limit-sort

Change-Id: I5abb1c7e2710af0be9dbd757608574945a26637e