Manually add pk for alembic_version table
We hit a problem that Neutron migrations cannot be applied
on galera cluster with ENFORCING mode as alembic_version table
missing primary key on the table and it is expected that all tables
will have primary keys.

alembic_version is created and managed by Alembic and fixes for this
on Alembic side are available since 0.8.10.
We cannot bump requirements for stable/newton but we can add
pk manually to allow new deployments with Newton code on Galera.

Change-Id: I543a1ee286bdf11ae35adb87125d044a351a2648
Closes-bug: #1655610