Protect against non-determinstic sort
This reverts commit e74b45d0eb923129902f1926aaddb51e6e2b8085.

This re-adds the non-deterministic sort logic but avoids using
boolean columns to work around the related bug.

Original commit msg:

With the switch to subquery loading, we need to eliminate the
possibility for non-deterministic results in the event of LIMIT
queries caused by non-unique SORT BY parameters.

This always adds on a set of unique keys to sort by when sorting
params are present. By placing them at the end, they take the lower
priority over the user-provided sort strings.

1. http://docs.sqlalchemy.org/en/latest/orm/loading_relationships.html#subqueryload-ordering
2. http://docs.sqlalchemy.org/en/latest/faq/ormconfiguration.html#faq-subqueryload-limit-sort

Related-Bug: #1656947
Change-Id: I503c65560d970427be93b39898f4e7acfc4d3e1f