Add IPv6 Prefix Delegation support for DVR
Unlike Legacy routers, DVR Edge Routers have their gateway
interfaces in the SNAT namespace as opposed to the router
namespace.  Added a new method to the router_info class,
get_gw_ns_name(), so callers can determine which namespace
the gateway device lives in.  This can then be over-ridden
in the DVR Edge router class.

The Prefix Delegation code will also now listen on the
update_router event from the l3-agent and reset the namespace
name if it changes.

Closes-Bug: #1541406
Change-Id: If6ada5027d0483fac7fc3ff935fee1edfc6e2759