tempest: Pin Ubuntu image to one version
The tests running with Ubuntu image used to pass and then suddenly all
of them stopped working. The reason was the public key was added to the
authorized_keys to root directory instead of ubuntu user directory. This
patch pins the used version to specific release date so we don't get
broken in the future with the same issue again.

Also this patch decreases amount of memory per guest to 512M as it
appears to work fine.

Change-Id: I93d7d4e72c1708017afecf22defbebc10bf6bcf7