Skip segment checks during network delete operations
These checks aren't necessary when triggered as part of a
network delete operation since any remaining subnets and ports
will be removed from the network. Additionally, they can race
with the auto cleanup methods for ports and subnets if segment
deletion is moved to BEFORE_DELETE where it should be.

Change-Id: I996f5b22f5179f5540760cc05c9f091bb3f3069b